requirements.txt
streamlit
google-generativeai
chromadb
uuid
